requirements.txt
anthropic>=0.40.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
python-dateutil>=2.8.2
pytest>=7.4.0
